Trait r3_port_arm_m::SysTickOptions

pub trait SysTickOptions {
    const FREQUENCY: u64;
    const FREQUENCY_DENOMINATOR: u64;
    const INTERRUPT_PRIORITY: InterruptPriority;
    const TICK_PERIOD: u32;
}

The configuration for the implementation of PortTimer based on SysTick (tickful).

Associated Constants

const FREQUENCY: u64

The numerator of the input clock frequency of SysTick.

const FREQUENCY_DENOMINATOR: u64

The denominator of the input clock frequency of SysTick. Defaults to 1.

const INTERRUPT_PRIORITY: InterruptPriority

The interrupt priority of the SysTick interrupt line. Defaults to 0xc0.

const TICK_PERIOD: u32

The period of ticks, measured in SysTick cycles. Must be in range 0..=0x1000000.

Defaults to (FREQUENCY / FREQUENCY_DENOMINATOR / 100).max(1).min(0x1000000) (100Hz).

Loading content...